% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_averages.R
\name{block_averages}
\alias{block_averages}
\title{Mean by blocks}
\usage{
block_averages(Y, rprime, grand_pixel = TRUE)
}
\arguments{
\item{Y}{a matrix}

\item{rprime}{the radius of blocks. For example, rprime = 1 implies a window of 3*3.}

\item{grand_pixel}{return a matrix with blocks of (2rprime+1)**2 or of 1 pixels only (by default)}
}
\value{
a matrix with averages by block
}
\description{
Mean by blocks
}
\examples{
block_averages(matrix(rnorm(25*25, 2,1), nrow = 25), rprime = 5)
}
