% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction.R
\name{direction}
\alias{direction}
\title{Title}
\usage{
direction(vector)
}
\arguments{
\item{vector}{a vector of 2 coordinates}
}
\value{
a vector which has the same direction as the one in parameter, but its coordinates are the smallest positive integer
}
\description{
Title
}
\examples{
direction(c(1,0))
}
