% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonPackage-package.R
\docType{package}
\name{geostatistics-package}
\alias{geostatistics}
\alias{geostatistics-package}
\title{geostatistics: Geostatistic maps}
\description{
This package considers randomly generated matrix as maps. It plots the maps and indicates the variance and the distribution of a realization. It does moving average to generate new maps which are spatially structured. It calculates the covariance between points on spatial structured maps. And it subsamples maps, with or without an average of the neighbored points.
}
\author{
\strong{Maintainer}: Juliette Chiapello \email{chiapellojuliette@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
