% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB.R
\name{VB}
\alias{VB}
\title{Variance Between}
\usage{
VB(Y, rprime)
}
\arguments{
\item{Y}{a matrix}

\item{rprime}{the radius of blocks. For example, rprime = 1 implies a window of 3*3.}
}
\value{
the variance between blocks
}
\description{
Variance Between
}
\examples{
VB(matrix(rnorm(15*20, 0, 1), nrow = 15), rprime = 2)
}
