% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VW.R
\name{VW}
\alias{VW}
\title{Variance Within}
\usage{
VW(Y, rprime)
}
\arguments{
\item{Y}{a matrix}

\item{rprime}{the radius of blocks. For example, rprime = 1 implies a window of 3*3.}
}
\value{
the mean of the variance inside blocks
}
\description{
Variance Within
}
\examples{
VW(matrix(rnorm(15*20, 0, 1), nrow = 15), rprime = 2)
}
